chatwithdatabase/requirement.txt
langchain
langchain_experimental
langchain_google_genai
python-dotenv
streamlit
tiktoken
faiss-cpu
protobuf
mysql-connector-python
pymysql
sentence.transformers
chromadb